# Fix performance of building row-level results