--- a/src/main/scala/com/amazon/deequ/VerificationResult.scala
+++ b/src/main/scala/com/amazon/deequ/VerificationResult.scala
@@ -97,10 +97,7 @@ object VerificationResult {
 
     val columnNamesToMetrics: Map[String, Column] = verificationResultToColumn(verificationResult)
 
-    val dataWithID = data.withColumn(UNIQUENESS_ID, monotonically_increasing_id())
-    columnNamesToMetrics.foldLeft(dataWithID)(
-      (dataWithID, newColumn: (String, Column)) =>
-        dataWithID.withColumn(newColumn._1, newColumn._2)).drop(UNIQUENESS_ID)
+    data.withColumns(columnNamesToMetrics)
   }
 
   def checkResultsAsJson(verificationResult: VerificationResult,