Fixes #576

Iteratively calling `withColumn` (singular) on a DataFrame causes performance issues when iterating over a large sequence of columns. (See issue for more details.) The code was iterating to add each `(name, column)` pair in a map to the DataFrame, so we can use [`withColumns`](https://spark.apache.org/docs/latest/api/scala/org/apache/spark/sql/Dataset.html#withColumns(colsMap:Map%5BString,org.apache.spark.sql.Column%5D):org.apache.spark.sql.DataFrame), which takes a `Map[String, Column]` as its parameter, as a drop-in replacement.

After running the performance test in the bug ticket, the performance is much better:
```
Gathering row-level results
51 columns in row level results
Duration was 67ms

Gathering row-level results
101 columns in row level results
Duration was 53ms

Gathering row-level results
151 columns in row level results
Duration was 49ms

Gathering row-level results
201 columns in row level results
Duration was 45ms

Gathering row-level results
251 columns in row level results
Duration was 48ms

Gathering row-level results
301 columns in row level results
Duration was 48ms

Gathering row-level results
351 columns in row level results
Duration was 41ms

Gathering row-level results
401 columns in row level results
Duration was 41ms
```

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.