@mentekid 2024-08-31
We don't have a set cadence but I think we can kick off a release some time
this week for this and some other changes that have accumulated since our
last one.

On Sat, Aug 31, 2024 at 13:58 Josh ***@***.***> wrote:

> Thanks @mentekid <https://github.com/mentekid>! By the way, any idea how
> soon a version with this can get published? We're trying to support some
> users with huge numbers of checks (hundreds to thousands) so I'm hoping to
> incorporate this in our application soon.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/awslabs/deequ/pull/577#issuecomment-2322994291>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABOFPXAAWEOXBHD3BTPHE53ZUH73XAVCNFSM6AAAAABNN6Q7NKVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDGMRSHE4TIMRZGE>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>