@marcantony 2024-09-06
Hey @mentekid, just wanted to ask about the release again. Could we maybe get something out early next week?