@marcantony 2024-09-10
Hey @mentekid, I just realized that `withColumns` was [added in Spark 3.3.0](https://spark.apache.org/docs/latest/api/scala/org/apache/spark/sql/Dataset.html#withColumns(colsMap:Map[String,org.apache.spark.sql.Column]):org.apache.spark.sql.DataFrame) so this is going to cause problems for the lower version builds. I'll make a PR today to replace it with a `select` instead.